First stab at a min-ified translation.
The results are interesting. Below, results28.txt is suspected to
correspond to a pre-min version, with CF = CF -> CF.  I say
"suspected" because I'm not completely sure, and badder and badderer
timing out surprises me.

The results for when the prelude CF = CF -> CF is not enabled (there
are still specialized versions for each function symbol) follow.  Note
that "FAILED WITH EXIT CODE 1" means Equinox found the theory

  T /\ ~ phi            (for T the axioms and phi the goal formula)

satisfiable/consistent. This is NOT the same as T /\ phi being
inconsistent, which would be bad; it simply means that T does not
entail phi.

  $ diff -U0 results/results28.txt results/38.txt
  --- results/results28.txt       2011-10-31 21:50:52.000000000 +0000
  +++ results/38.txt      2011-11-02 20:39:22.000000000 +0000
  @@ -5 +9 @@
  -./ok/add.hs:                  passed
  +./ok/add.hs:                  FAILED WITH EXIT CODE 1
  @@ -8 +12 @@
  -./ok/copy.hs:                 passed
  +./ok/copy.hs:                 FAILED WITH EXIT CODE 1
  @@ -10 +14 @@
  -./ok/implies.hs:              TIMED OUT
  +./ok/implies.hs:              FAILED WITH EXIT CODE 1
  @@ -12 +16 @@
  -./ok/map.hs:                  passed
  +./ok/map.hs:                  FAILED WITH EXIT CODE 1
  @@ -16 +20 @@
  -./ok/odd.hs:                  passed
  +./ok/odd.hs:                  FAILED WITH EXIT CODE 1
  @@ -25,2 +29,2 @@
  -./diverges/badderer.hs:       timed out
  -./diverges/badder.hs:         timed out
  +./diverges/badderer.hs:       FAILED WITH EXIT CODE 1
  +./diverges/badder.hs:         FAILED WITH EXIT CODE 1
  @@ -30,2 +34,2 @@
  -./diverges/extrasimple.hs:    timed out
  -./diverges/list.hs:           timed out
  +./diverges/extrasimple.hs:    FAILED WITH EXIT CODE 1
  +./diverges/list.hs:           FAILED WITH EXIT CODE 1

Now, 'add', 'implies', 'map', and 'odd' all depend on 'CF = CF -> CF',
so it's not surprising that T does not entail phi.  But, in addition,
'implies' no longer times out. And, finding the theory consistent is
much better than timing out for these.  Even better, 'badder',
'badderer', and 'extrasimple' are all false, i.e. T /\ phi is
inconsistent for these, so Equinox finding T /\ ~ phi consistent is
the best we could hope for.  And finally, 'list' doesn't even type
check :P

With the prelude 'CF = CF -> CF' enabled, things don't go well:

  $ diff -U0 results/38.txt results/39.txt
  --- results/38.txt      2011-11-02 20:39:22.000000000 +0000
  +++ results/39.txt      2011-11-02 20:59:18.000000000 +0000
  @@ -9 +9 @@
  -./ok/add.hs:                  FAILED WITH EXIT CODE 1
  +./ok/add.hs:                  TIMED OUT
  @@ -12 +12 @@
  -./ok/copy.hs:                 FAILED WITH EXIT CODE 1
  +./ok/copy.hs:                 TIMED OUT
  @@ -14 +14 @@
  -./ok/implies.hs:              FAILED WITH EXIT CODE 1
  +./ok/implies.hs:              TIMED OUT
  @@ -16 +16 @@
  -./ok/map.hs:                  FAILED WITH EXIT CODE 1
  +./ok/map.hs:                  TIMED OUT
  @@ -20,2 +20,2 @@
  -./ok/odd.hs:                  FAILED WITH EXIT CODE 1
  -./ok/reverse.hs:              passed
  +./ok/odd.hs:                  TIMED OUT
  +./ok/reverse.hs:              TIMED OUT
  @@ -23 +23 @@
  -./ok/test2.hs:                passed
  +./ok/test2.hs:                TIMED OUT
  @@ -29,2 +29,2 @@
  -./diverges/badderer.hs:       FAILED WITH EXIT CODE 1
  -./diverges/badder.hs:         FAILED WITH EXIT CODE 1
  +./diverges/badderer.hs:       timed out
  +./diverges/badder.hs:         timed out
  @@ -34,2 +34,2 @@
  -./diverges/extrasimple.hs:    FAILED WITH EXIT CODE 1
  -./diverges/list.hs:           FAILED WITH EXIT CODE 1
  +./diverges/extrasimple.hs:    timed out
  +./diverges/list.hs:           timed out

But withq both the prelude and function specific 'CF = CF -> CF'
removed we get

  $ diff -U0 results/38.txt results/40.txt
  --- results/38.txt      2011-11-02 20:39:22.000000000 +0000
  +++ results/40.txt      2011-11-02 21:18:09.000000000 +0000
  @@ -17,3 +17,3 @@
  -./ok/multgt.hs:               TIMED OUT
  -./ok/mult.hs:                 TIMED OUT
  -./ok/nateqsymmetry.hs:        TIMED OUT
  +./ok/multgt.hs:               FAILED WITH EXIT CODE 1
  +./ok/mult.hs:                 FAILED WITH EXIT CODE 1
  +./ok/nateqsymmetry.hs:        passed
  @@ -24 +24 @@
  -./ok/test.hs:                 TIMED OUT
  +./ok/test.hs:                 FAILED WITH EXIT CODE 1

where 'nateqsymmetry' doesn't type check (or even define all the
functions it uses :P), 'test' uses 'CF = CF -> CF', and 'mult' and
'multgt' use recursions for which there is no 'min' info.  This last
part is of interest: e.g. for

  f x = case x of
          K -> g x
  g x = e

the formulas generated for 'f' don't give 'min(g x)' in the 'K'
branch, and so the proof should get stuck.  Fix that next ...