Paper: updated expression langauge; limit 'full' application to axioms.
Added 'UNR' and function symbols to the expression language, and made
most examples use partial application.  Axioms still use full application.
The point is that full application is only relevant as an optimization
of generated formulas to make the theorem prover more easily recognize
when a formula applies to a particular term. E.g., we think it's easier
for the prover to match on 'f(x1,...,xn)' then 'f `app` x1 `app` ... `app` xn'.