Fixed phi_cf per discussion with Dimitrios.
See paper.pdf or contracts/todo.org for discussion.  In short, removed

  min(f x) -> min(x)

and changed phi_cf to, roughly,

  min(K xs) -> cf(K xs) -> forall x. cf(x)
  /\
  [[K in CF^n -> CF]]-

for n the arity of K.  In addition to providing a better solution to
the problem of propagating interest into arguments when proving CF for
constructor applications, it also improves test results:

  $ diff -U0 ./egs/results/44.20s-equinox.txt ./egs/results/45.20s-equinox.txt
  --- ./egs/results/44.20s-equinox.txt    2011-11-06 21:49:36.000000000 +0000
  +++ ./egs/results/45.20s-equinox.txt    2011-11-07 17:51:46.000000000 +0000
  @@ -25 +25 @@
  -./egs/no/add-and-mult-nonZero.hs:                 timed out
  +./egs/no/add-and-mult-nonZero.hs:                 FAILED WITH EXIT CODE 1
  @@ -36 +36 @@
  -./egs/no/nat-max.hs:                              timed out
  +./egs/no/nat-max.hs:                              FAILED WITH EXIT CODE 1

Both of those examples are false, so this is the best we can hope for.