Integrated the 'order to check things' code and fixed multiple/mutual contracts.
This gets us back to where we were before the switch to haskell syntax,
in terms of performance, with the addition of multiple/mutual contract
support.  By "mutual" I mean when two mutually recursive functions are
checked.

The current diff is:

  $ diff -U0 results/{46,55}*
  --- results/46.20s-equinox.txt  2011-11-08 15:20:04.000000000 +0000
  +++ results/55.20s-equinox.txt  2011-11-18 19:32:28.000000000 +0000
  @@ -17,0 +18 @@
  +./yes/gt-inc.hs:                                  TIMED OUT
  @@ -23 +24 @@
  -./yes/prediate-on-constant-function.hs:           passed
  +./yes/predicate-on-constant-function.hs:          passed
  @@ -33,0 +35,2 @@
  +./no/mult-gt.hs:                                  FAILED WITH EXIT CODE 1
  +./no/mutual-recursion-bug.hs:                     FAILED WITH EXIT CODE 1

I.e., it's as it was before, modulo some new tests and a typo fix.

This diff includes a lot of whitespace only adjustments.  Use

  $ git diff -b

to see a whitespace ignored diff.