Fixed the bug in my new translation.
I had to rewrite the definition translation (dTrans) to support
the nested case expressions.  Along the way I made a small
"irrelevant" change, that turned out to be very wrong :P
I added a comment explaining the mistake, but in short, I
assumed that (A/\B)\/C -> (A/\B)\/(~A/\C) :P This allowed
me to simplify a formula ... and make it wrong.