Switched to a proper command line option handler: CmdArgs.
The old ad-hoc setup was annoying, e.g. you had to list the file name
before the options, the help messages had to be formatted manually,
and all the option handling had to be coded.  Now we git everything
for free :) I also changed the theoremprover option to be an Enum, but
that was a waste of time (I was trying to do something better, but got
stuck because you can't derive Data for arrow types, so I adapted what
I had started, instead of reverting, which would probably been
quicker.