10 character fix for a 5 hour bug :P
I placed the min to restrictively on phi_project, resulting in the inability
to prove that constructors are invertible, unless you're already interested
in their projection.  Figured this out because I tried to do a proof on
paper that we expected the theorem prover to do, and then got stuck
when i needed to prove 'Zero = X_' from 'Succ Zero = Succ X_'
and 'min(Succ Zero)'.

Ouch.