Replaced the 'data MetaT a = ...', 'type T = MetaT E' stuff with GP.
I wrote a generic-programming function 'gmap' that applys a
transformation to all maximal subterms of the appropriate type.  For
example, you can 'gmap (reverse::[T]->[T]) e' to reverse all maximal
lists in 'e'.  Here "maximal" means "not the tail of a cons".  Note
that if you try

  everywhere (mkT (reverse::[T]->[T])) e

you'll reverse the tail before reversing the list, making a mess. E.g.

  everywhere (mkT (reverse::[T]->[T])) [[1,2,3]] == [[2,3,1]]

The 'gfmap' is not as precise as 'fmap', in that you can't distinguish
between different subterms of the same type.  E.g.

  gfmap ((+1)::Int->Int) ((0,0)::(Int,Int)) = (1,1)

whereas

   fmap  (+1)             (0,0)             = (0,1)

On the other hand, fmap can't transform the first component, even if
you want it to.