Fixed a bug in dependency graph generation for repeated imports.
This bug was identified by 'assert' statements added much earlier,
so I don't think it ever affected any results.  The problem was
that importing the same lib multiple times resulted in a the imported
names appearing more than once in the dependency graph, which
violates an (implicit ?) precondition of the SCC lib we use.

The multiple import, which is allowed, happened when I moved
lem_eqNat_trans into a library.